feat(sheets): 子物品浮点倍数 + 父子排序 + 三端 UX 修正
响应实测 9 项（网页 5 + MCDR 4）。后端：迁移 0013（qty_per_unit INTEGER→NUMERIC(10,2)、CHECK ≥1→>0），子行派生 need_qty=ceil(qty_per_unit×父.need_qty) 保持整数，sort_sheet_rows 父子感知修复『子行排到父行上方』排序 bug，全链路 float 适配。前端 SheetEditor 改 el-table tree mode（删游离子表+手动展开状态机+重复操作列，净 -190 行），文字按钮，倍数浮点列，父行『添加子物品』popover，数字输入加 placeholder。MCDR：[改ID]→[改]、[+子]→[子]，addsub 改手持扫描复用 scanner.read_held_item，qty_per_unit Float 节点，模板『每件×N』去尾零。测试：后端 167（含浮点 ceil + 父子排序新增用例）/MCDR 219/前端 build 绿。

+++ McdrPlugin/htcmc_auth/htcmc_auth/sheet_client.py
@@ -160,7 +160,7 @@ def upsert_row(
     registry_id: Optional[str] = None,
     row_id: Optional[int] = None,
     parent_row_id: Optional[int] = None,
-    qty_per_unit: Optional[int] = None,
+    qty_per_unit: Optional[float] = None,
 ) -> SheetOutcome:
     """PUT /sheets/{sheet_id}/rows 单端点按 row_id 分流（issue #20）→ RowDetail。
